+++ script/startup.sh
@@ -222,9 +222,6 @@ function read_vendor_param() {
 	        # Without a Facet Chrome app, we are likely running Vendorless BCLD
 	        list_item 'BCLD App not found, setting to: VENDORLESS BCLD'
 		    export BCLD_VENDOR='vendorless'
-		    
-		    # Vendorless BCLD should swap mouse buttons to enable use of tabs by default
-		    export BCLD_SWAP=1
 		fi
 	else
 		# Display used BCLD_VENDOR parameter