+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 ## 2024
 
 ## April
-* `2024-04-11 16:35:55` : [OAC-905] Added new BCLD parameter: [bcld.afname.swap](https://github.com/jbbletterman/bcld/wiki/Client%20Configuration#general)
-    - Can be used for swapping mouse buttons M2 and M3
-    - Allows for usage of tabs in Vendorless BCLD
+* `2024-04-11 16:35:55` : [OAC-905] Swapping mouse buttons 2 and 3 for Vendorless BCLD by default in [Xconfigure](./script/Xconfigure.sh#L99)
+    - Mouse button 2 does not work in Vendorless BCLD
+    - Swapping mouse buttons allows for usage of tabs on laptops
 * `2024-04-10 17:11:10` : [OAC-1031] Enhanced `BCLD_CERTs` and `BCLD_KEYs` methods in [BCLD TEST package](./test/bcld_test.sh) to also include checks for client certificate and key
 * `2024-04-04 17:24:11` : Migrated system requirements, USB configurations and tools from [README](./README.md) to [BCLD Wiki](https://www.github.com/jbbletterman/bcld/wiki)